Dry run added.

+++ shutil/parallel-tree-copy.py
@@ -30,6 +30,7 @@ from typing import Optional
 
 MAX_WORKERS = 16
 
+
 class InvalidConfigurationError(Exception):
     ...
 
@@ -40,6 +41,7 @@ class Configuration:
     destination_path: str
     regex_filter: Optional[str]
     workers: Optional[int]
+    dry_run: bool
 
     def __post_init__(self) -> None:
         if not Path(self.source_path).is_dir():
@@ -71,7 +73,7 @@ class Summary:
 class Stopwatch:
 
     def __init__(self) -> None:
-        self.start_time = perf_counter()
+        self._start_time = perf_counter()
 
     @staticmethod
     def start() -> Stopwatch:
@@ -79,7 +81,7 @@ class Stopwatch:
 
 
     def elapsed_time_millis(self) -> int:
-        duration = perf_counter() - self.start_time
+        duration = perf_counter() - self._start_time
         return int(1000 * duration)
 
 
@@ -94,7 +96,11 @@ not matching the regex filter will be skipped. The filter is case-insensitive.
 
 
 def create_cmd_line_args_parser() -> ArgumentParser:
-    parser = ArgumentParser(description="Parallel Copying of Directory Structures", formatter_class=RawTextHelpFormatter, epilog=epilog())
+    parser = ArgumentParser(
+        description="Parallel Copying of Directory Structures",
+        formatter_class=RawTextHelpFormatter,
+        epilog=epilog()
+    )
 
     # positional mandatory arguments
     parser.add_argument("source_path",
@@ -114,6 +120,12 @@ def create_cmd_line_args_parser() -> ArgumentParser:
         default=4,
         help=f"optional number of worker threads to be used (default = 4, max = {MAX_WORKERS})",
         type=int)
+    parser.add_argument("-d", "--dry-run",
+        dest="dry_run",
+        default=False,
+        action="store_true",
+        help="if specified, the script will not copy any files (it will only print what would be copied)"
+    )
 
     return parser
 
@@ -125,7 +137,8 @@ def parse_cmd_line_args() -> Configuration:
         source_path=params.source_path,
         destination_path=params.destination_path,
         regex_filter=params.regex_filter,
-        workers=params.workers
+        workers=params.workers,
+        dry_run=params.dry_run,
     )
 
 
@@ -137,8 +150,7 @@ def get_sorted_subdirs(config: Configuration) -> tuple[str, ...]:
     for path in Path(config.source_path).iterdir():
         if not path.is_dir():
             continue
-        if config.regex_filter is not None:
-            if regex is not None and not regex.match(path.name):
+        if regex is not None and not regex.match(path.name):
                 ignored_count += 1
                 continue
         result.append(str(path))
@@ -152,11 +164,12 @@ def get_sorted_subdirs(config: Configuration) -> tuple[str, ...]:
     return tuple(sorted(result))
 
 
-def copy_subdir(request: CopyRequest) -> CopyResult:
-    print(f"Going to copy {request.source} to {request.destination}")
+def copy_subdir(request: CopyRequest, dry_run: bool) -> CopyResult:
+    print(f"Going to copy '{request.source}' to '{request.destination}")
     stopwatch = Stopwatch.start()
     try:
-        copytree(request.source, request.destination, dirs_exist_ok=True)
+        if not dry_run:
+            copytree(request.source, request.destination, dirs_exist_ok=True)
         duration_millis = stopwatch.elapsed_time_millis()
         return CopyResult(
             request=request,
@@ -182,7 +195,7 @@ def copy_subdirs(config: Configuration, source_list: tuple[str, ...]) -> Summary
                 source=source,
                 destination=str(destination)
             )
-            future = executor.submit(copy_subdir, request)
+            future = executor.submit(copy_subdir, request, config.dry_run)
             future_list.append(future)
         
         success_count = 0
@@ -191,10 +204,10 @@ def copy_subdirs(config: Configuration, source_list: tuple[str, ...]) -> Summary
             result = future.result()
             if result.exception is None:
                 formatted_duration = format_duration(result.duration_millis)
-                print(f"Successfully copied {result.request.source} to {result.request.destination} in {result.duration_millis} ms ({formatted_duration})")
+                print(f"Successfully copied '{result.request.source}' to '{result.request.destination}' in {result.duration_millis} ms ({formatted_duration})")
                 success_count += 1
             else:
-                print(f"Failed to copy {result.request.source} to {result.request.destination}")
+                print(f"Failed to copy '{result.request.source}' to '{result.request.destination}'")
                 print(str(result.exception))
                 failure_count += 1
 
@@ -221,6 +234,7 @@ def print_summary(config: Configuration, summary: Summary) -> None:
     print(f"Overall duration:                   {summary.overall_duration_millis} ms ({formatted_duration})")
     print(f"Successfully copied subdirectories: {summary.success_count}")
     print(f"Failed subdirectories:              {summary.failure_count}")
+    print(f"Dry run:                            {'yes' if config.dry_run else 'no'}")
     print()
